requirements.txt
absl-py==1.3.0
appnope==0.1.3
asttokens==2.2.0
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4d2o0arbv6/croot/certifi_1665076671099/work/certifi
charset-normalizer==2.1.1
cloudpickle==2.2.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
dm-tree==0.1.7
entrypoints==0.4
executing==1.2.0
flatbuffers==22.11.23
fonttools==4.38.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.7
jupyter_core==5.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.1rc2
libclang==14.0.6
littleutils==0.2.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
outdated==0.2.2
packaging==21.3
pandas==1.5.2
pandas-flavor==0.3.0
paranoid-scientist==0.2.2
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pingouin==0.5.2
platformdirs==2.5.4
prompt-toolkit==3.0.33
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyddm==0.6.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.9.3
seaborn==0.11.2
six==1.16.0
stack-data==0.6.2
statannotations==0.5.0
statsmodels==0.13.5
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.18.0
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-probability==0.18.0
termcolor==2.1.1
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
xarray==2022.11.0
zipp==3.11.0
